Use hook instead of helper function to get included product description map.
